+++ src/main/java/com/dd/ai_smart_course/service/impl/QAImpl.java
@@ -27,7 +27,7 @@ public class QAImpl implements QAService {
     }
 
     @Override
-    public QA getQAById(int id) {
+    public QA getQAById(long id) {
         return qaMapper.getQAById(id);
     }
 
@@ -61,7 +61,7 @@ public class QAImpl implements QAService {
     }
 
     @Override
-    public int deleteQA(int id) {
+    public int deleteQA(long id) {
         return qaMapper.deleteQA(id);
     }
 
@@ -84,7 +84,7 @@ public class QAImpl implements QAService {
     //根据QAId检查qa是否存在
     //针对 update
     @Override
-    public void checkQAExists(Integer qaId) throws BusinessException{
+    public void checkQAExists(Long qaId) throws BusinessException{
         if(qaId == null){
             throw new BusinessException(ErrorCode.QA_ID_NULL);
         }
